refactor(light): split multipole module + add ag.lp_linear variants
Refactors the multipole light profile module added in #420:

- autogalaxy/profiles/light/standard/multipole.py is split into per-class
  modules:
  - _multipole_mixin.py — private, holds the shared
    _LightProfileMultipoleMixin
  - sersic_multipole.py — SersicMultipole
  - gaussian_multipole.py — GaussianMultipole
- test_autogalaxy/profiles/light/standard/test_multipole.py is split into
  test_sersic_multipole.py + test_gaussian_multipole.py
- standard/__init__.py updated to the two new imports

Adds linear-light-profile counterparts:

- autogalaxy/profiles/light/linear/sersic_multipole.py
  ag.lp_linear.SersicMultipole(lp.SersicMultipole, LightProfileLinear)
  hardcodes intensity=1.0 (solved via inversion) and drops intensity from
  the constructor signature.
- autogalaxy/profiles/light/linear/gaussian_multipole.py — same pattern.
- linear/__init__.py exports both.
- test_autogalaxy/profiles/light/linear/test_sersic_multipole.py and
  test_gaussian_multipole.py cover: class identity (LightProfileLinear +
  the standard subclass), intensity-not-a-constructor-parameter,
  multipole_*_comps propagation, parity with the linear base profile at
  zero perturbation, and image parity with the standard variant at
  intensity=1.0.

Docs:
- docs/api/light.rst Linear [ag.lp_linear] section now lists
  SersicMultipole and GaussianMultipole.

895 unit tests pass. Follows up #418.
