+++ autogalaxy/profiles/light/linear/__init__.py
@@ -6,6 +6,8 @@ from .exponential import Exponential, ExponentialSph
 from .dev_vaucouleurs import DevVaucouleurs, DevVaucouleursSph
 from .sersic_core import SersicCore, SersicCoreSph
 from .exponential_core import ExponentialCore, ExponentialCoreSph
+from .sersic_multipole import SersicMultipole
+from .gaussian_multipole import GaussianMultipole
 from .shapelets.polar import ShapeletPolarSph, ShapeletPolar
 from .shapelets.cartesian import ShapeletCartesianSph, ShapeletCartesian
 from .shapelets.exponential import ShapeletExponentialSph, ShapeletExponential
