+++ autogalaxy/profiles/light/standard/__init__.py
@@ -13,7 +13,8 @@ from .eff import (
     ElsonFreeFall,
     ElsonFreeFallSph,
 )
-from .multipole import SersicMultipole, GaussianMultipole
+from .sersic_multipole import SersicMultipole
+from .gaussian_multipole import GaussianMultipole
 from .shapelets.polar import ShapeletPolarSph, ShapeletPolar
 from .shapelets.cartesian import ShapeletCartesianSph, ShapeletCartesian
 from .shapelets.exponential import ShapeletExponentialSph, ShapeletExponential
